[Gold V] Title: 이진 검색 트리, Time: 0 ms, Memory: 2604 KB -BaekjoonHub

+++ 백준/Gold/5639. 이진 검색 트리/이진 검색 트리.cc
@@ -3,7 +3,7 @@ using namespace std;
 
 struct Node {
 	int num;
-	int left, right;
+	short left, right;
 	Node() { left = right = 1; }
 };
 int arr[10000];
